requirements.txt
asgiref==3.6.0
black==21.9b0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==3.2.7
django-autoslug==1.9.8
django-cors-headers==3.14.0
django-countries==7.2.1
django-filter==21.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==3.9.2
gunicorn==20.1.0
idna==3.4
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
pathspec==0.11.1
phonenumbers==8.12.33
Pillow==8.3.2
platformdirs==3.1.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.6.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2022.7.1
regex==2023.3.22
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.3
tomli==1.2.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
